skynet/requirements.txt
# SKYNET - Python Dependencies

# Core
aiohttp>=3.10,<4.0
aiosqlite>=0.20.0,<1.0

# WebSocket
websockets>=14.0,<15.0

# Telegram
python-telegram-bot>=21.0,<22.0

# AI Providers
google-genai>=1.0
groq>=0.8.0
openai>=1.0
anthropic>=0.30.0

# Queue (Celery + Redis)
celery>=5.3.0
redis>=5.0.0

# AWS S3
boto3>=1.34.0

# HTTP Client
httpx>=0.27.0

# YAML config
pyyaml>=6.0.0
